+++ src/dipdup/dipdup.py
@@ -171,15 +171,6 @@ class IndexDispatcher:
                 self._ctx.config.contracts[contract.name] = contract_config
         self._ctx.config.initialize(skip_imports=True)
 
-    async def _subscribe_to_datasource_events(self) -> None:
-        for datasource in self._ctx.datasources.values():
-            if not isinstance(datasource, IndexDatasource):
-                continue
-            datasource.on_head(self._on_head)
-            datasource.on_operations(self._on_operations)
-            datasource.on_big_maps(self._on_big_maps)
-            datasource.on_rollback(self._on_rollback)
-
     async def _load_index_states(self) -> None:
         if self._indexes:
             raise RuntimeError('Index states are already loaded')
@@ -222,6 +213,22 @@ class IndexDispatcher:
 
         tasks = (create_task(_process(index_state)) for index_state in await IndexState.all())
         await gather(*tasks)
+
+    async def _subscribe_to_datasource_events(self) -> None:
+        for datasource in self._ctx.datasources.values():
+            if not isinstance(datasource, IndexDatasource):
+                continue
+            datasource.call_on_disconnected(self._on_disconnected)
+            datasource.call_on_head(self._on_head)
+            datasource.call_on_operations(self._on_operations)
+            datasource.call_on_big_maps(self._on_big_maps)
+            datasource.call_on_rollback(self._on_rollback)
+
+    async def _on_disconnected(self) -> None:
+        # NOTE: Invalidate realtime queues; sync level will be reset
+        self._logger.info('Datasource disconnected, dropping realtime queues')
+        for index in self._indexes.values():
+            index._queue.clear()
 
     async def _on_head(self, datasource: IndexDatasource, head: HeadBlockData) -> None:
         # NOTE: Do not await query results - blocked database connection may cause Websocket timeout.