+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixed
 
 * index: Fixes crash with `RuntimeError` after continuous realtime connection loss.
+* index: Fixed `OperationIndexConfig.types` field being partially ignored.
+* tzkt: Fixed `origination` subscription missing when `merge_subscriptions` flag is set.
 
 ## 5.1.3 - 2022-05-26