+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [unreleased]
 
+### Fixed
+
+* index: Fixed `OperationIndexConfig.types` field being partially ignored.
+* tzkt: Fixed `origination` subscription missing when `merge_subscriptions` flag is set.
+
 ### Performance
 
 * ci: Decrease the size of generic and `-pytezos` Docker images by 11% and 16%, respectively.