+++ CHANGELOG.md
@@ -4,9 +4,14 @@
 
 ### Fixed
 
-* index: Fixes crash with `RuntimeError` after continuous realtime connection loss.
+* index: Allow mixing oneshot and regular indexes in a single config.
+* index: Fixed crash with `RuntimeError` after continuous realtime connection loss.
 * index: Fixed `OperationIndexConfig.types` field being partially ignored.
 * tzkt: Fixed `origination` subscription missing when `merge_subscriptions` flag is set.
+
+### Performance
+
+* ci: Decrease the size of generic and `-pytezos` Docker images by 11% and 16%, respectively.
 
 ## 5.1.3 - 2022-05-26